backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "layoutir>=1.0.3",
    "fastmcp>=2.0.0",
    "supabase>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
